project/requirements.txt
asgiref==3.4.1
black==21.12b0
boto3==1.20.25
botocore==1.23.25
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-debug-toolbar==2.2
django-extensions==3.1.5
django-filter==21.1
django-rest-authtoken==2.1.4
django-storages==1.12.3
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==9.8.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
PyJWT==1.7.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
